requirements.txt
# Core web framework & async runtime
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Async PostgreSQL driver & ORM
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0,<3.0.0

# Pydantic v2
pydantic>=2.9.0,<3.0.0
pydantic-settings>=2.5.0

# Redis (Critical for Thread Memory)
redis>=5.0.0

# Rate limiting
slowapi>=0.1.9

# Auth
pyjwt>=2.9.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
alembic>=1.13.0


# Environment & Utilities
python-dotenv>=1.0.0
httpx>=0.27.0           # Required for bot_runner.py API calls
PyYAML>=6.0.1           # Required for bot_loader.py (Fixes 'No module named yaml')
openai>=1.40.0          # Required for llm_client.py (Analyst Brain)

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0

# CLI / formatting
rich>=13.7.0